+++ src/screens/hardware-wallets/PairHardwareWalletErrorSheet.tsx
