+++ src/screens/hardware-wallets/PairHardwareWalletIntroSheet.tsx
