+++ src/screens/hardware-wallets/PairHardwareWalletSearchSheet.tsx
